// src/services/schemas/modelsChatResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 */

export interface ModelsChatResponse {
  /** channel */
  channelId?: number;
  createdAt?: string;
  fileName?: string;
  id?: number;
  /** Relate to type message */
  receiverId?: number;
  senderId?: number;
  senderName?: string;
  text?: string;
  /** "direct" | "channel" */
  type?: string;
  url?: string;
}